+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:shine_app/product_display.dart';
+import 'package:shine_app/screens/product_display/product_display.dart';
 import './gallery.dart';
 import 'models/product.dart';
 
@@ -41,12 +41,9 @@ class MyApp extends StatelessWidget {
       switch (settings.name) {
         case '/product':
           if (arguments is Product) {
-            // the details page for one specific user
             return MaterialPageRoute(builder: (context) => ProductDisplay(arguments));
           }
-          else {
-            break;
-          }
+          break;
         default:
           return MaterialPageRoute(builder: (context) => MyHomePage(title: "SHEIN"));
       }
